refactor plotter module
